requirements.txt
Django==1.7
Pillow==2.5.3
Pygments==1.6
argparse==1.2.1
certifi==14.05.14
django-filter==0.9.2
django-thumbs==0.4
djangorestframework==2.4.2
oauthlib==0.6.3
python-openid==2.2.5
python-social-auth==0.1.26
requests==2.4.0
requests-oauthlib==0.4.1
six==1.7.3
wsgiref==0.1.2
